删除多余文件，修复parse警告

+++ c++/SimpleDPP.hpp
@@ -155,7 +155,7 @@ public:
 
     void parse(const std::vector<sdp_byte> &data)
     {
-        for (int i = 0; i < data.size(); i++)
+        for (size_t i = 0; i < data.size(); i++)
         {
             parse(data[i]);
         }
